Add GitHub Action for Code Linting
This pull request introduces a GitHub Action workflow to automate the linting
of various code types in the repo on each push and pull request.

It add:
- lint.yml to define the workflow
- eslint.config.js to manage the rules for the js linting

The workflow runs the following linters:
- JavaScript: Uses ESLint to check for issues in JavaScript files.
- HTML: Uses HTMLHint to lint HTML files.
- CSS: Uses Stylelint to ensure proper CSS formatting.
- Shell scripts: Uses ShellCheck to lint shell scripts.
- C++: Uses Clang-Tidy to check C++ code for style and potential errors.

By adding this workflow, we ensure consistent code quality and formatting
across multiple languages in the repository.